refactor(models): flatten StandardizationTemplate packages per spec (Rule 0007 A) and export Blueprint classes (E)
