--- src/armodel/models/M2/AUTOSARTemplates/CommonStructure/StandardizationTemplate/BlueprintMapping/__init__.py
@@ -1,3 +1,0 @@
-from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.BlueprintMapping.BlueprintMappingSet import BlueprintMappingSet
-
-__all__ = ["BlueprintMappingSet"]
