+++ src/armodel/models/M2/AUTOSARTemplates/GenericStructure/AbstractStructure.py
@@ -5,9 +5,10 @@ in the GenericStructure module.
 
 from abc import ABC
 from typing import List, Optional
-from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import Identifiable, PackageableElement
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import Identifiable
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
+from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.AbstractBlueprintStructure import AtpBlueprintable
 
 
 class AtpInstanceRef(ARObject, ABC):
@@ -85,32 +86,6 @@ class AtpInstanceRef(ARObject, ABC):
         if value is not None:
             self.atpTargetRef = value
         return self
-
-
-class AtpBlueprintable(PackageableElement, ABC):
-    """
-    Abstract base class for AUTOSAR Template (ATP) blueprintable elements.
-
-    AtpBlueprintable represents elements that can be used as blueprints in the AUTOSAR
-    template system. These elements provide reusable definitions that can be instantiated
-    or referenced in the model.
-
-    This class extends Identifiable with blueprint-specific functionality for managing
-    template-based elements in AUTOSAR models.
-
-    Note:
-        This is an abstract class and cannot be instantiated directly.
-        Concrete implementations include BswModuleEntry, CompuMethod, DataConstr,
-        and other blueprintable AUTOSAR elements.
-    """
-
-    # AtpBlueprintable method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self, parent: ARObject, short_name: str):
-        if type(self) is AtpBlueprintable:
-            raise TypeError("AtpBlueprintable is an abstract class.")
-        super().__init__(parent, short_name)
 
 
 class AtpFeature(Identifiable, ABC):
@@ -271,33 +246,3 @@ class AtpStructureElement(AtpBlueprintable, ABC):
         if type(self) is AtpStructureElement:
             raise TypeError("AtpStructureElement is an abstract class.")
         super().__init__(parent, short_name)
-
-
-class AtpBlueprintMapping(ARObject, ABC):
-    """
-    Abstract base class for AUTOSAR Template (ATP) blueprint mapping elements.
-
-    AtpBlueprintMapping represents mapping elements in the AUTOSAR system that
-    define relationships between blueprints and their implementations or instances.
-    Mappings provide the mechanism to connect abstract blueprint definitions
-    with concrete implementations.
-
-    This class extends ARObject with mapping-specific functionality for managing
-    blueprint mapping relationships.
-
-    Note:
-        This is an abstract class and cannot be instantiated directly.
-        AtpBlueprintMapping is the parent of various AUTOSAR mapping elements:
-        - BlueprintMapping (generic blueprint to implementation mapping)
-
-    Attributes:
-        Inherits all attributes from ARObject including uuid and adminData.
-    """
-
-    # AtpBlueprintMapping method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self):
-        if type(self) is AtpBlueprintMapping:
-            raise TypeError("AtpBlueprintMapping is an abstract class.")
-        super().__init__()
