+++ tests/test_armodel/models/M2/AUTOSARTemplates/CommonStructure/StandardizationTemplate/test_BlueprintMappingSet.py
@@ -3,7 +3,7 @@ This module contains tests for the BlueprintMappingSet class in the
 AUTOSAR CommonStructure.StandardizationTemplate.BlueprintMapping module.
 """
 
-from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.BlueprintMapping.BlueprintMappingSet import (
+from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.BlueprintMapping import (
     BlueprintMappingSet,
 )
