+++ tests/test_armodel/models/M2/AUTOSARTemplates/GenericStructure/test_VariantHandling.py
@@ -2,7 +2,7 @@
 Tests for GenericStructure VariantHandling model classes.
 """
 
-from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.BlueprintGenerator.BlueprintGenerator import (
+from armodel.models.M2.AUTOSARTemplates.CommonStructure.StandardizationTemplate.BlueprintGenerator import (
     BlueprintGenerator,
 )
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ARPackage import (
